Remove global resources var The `provider` package now has a function that will expose all the resources from all subcategories of the provider This list is used to: - Used in the import docs generation command - Sent to both provider servers - Used in the tf code generation command
